requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
tensorflow==2.13.0
pillow==9.5.0
numpy==1.24.2